fix: reorganize endpoints
