+++ src/api.ts
@@ -1,50 +1,11 @@
 import express from 'express';
-import { rpcError, rpcSuccess, storageEngine } from './helpers/utils';
 import log from './helpers/log';
-import { queues } from './lib/queue';
+import { rpcError, storageEngine } from './helpers/utils';
 import getModerationList from './lib/moderationList';
-import { name, version } from '../package.json';
 import VotesReport from './lib/votesReport';
+import { queues } from './lib/queue';
 
 const router = express.Router();
-
-router.get('/', (req, res) => {
-  const commit = process.env.COMMIT_HASH || '';
-  const v = commit ? `${version}#${commit.substr(0, 7)}` : version;
-  return res.json({
-    name,
-    version: v
-  });
-});
-
-router.post('/votes/generate', async (req, res) => {
-  log.info(`[http] POST /votes/generate`);
-
-  const body = req.body || {};
-  const event = body.event.toString();
-  const id = body.id.toString().replace('proposal/', '');
-
-  if (req.headers['authenticate'] !== process.env.WEBHOOK_AUTH_TOKEN?.toString()) {
-    return rpcError(res, 'UNAUTHORIZED', id);
-  }
-
-  if (!event || !id) {
-    return rpcError(res, 'Invalid Request', id);
-  }
-
-  if (event !== 'proposal/end') {
-    return rpcSuccess(res, 'Event skipped', id);
-  }
-
-  try {
-    await new VotesReport(id, storageEngine(process.env.VOTE_REPORT_SUBDIR)).canBeCached();
-    queues.add(id);
-    return rpcSuccess(res, 'Cache file generation queued', id);
-  } catch (e) {
-    log.error(e);
-    return rpcError(res, 'INTERNAL_ERROR', id);
-  }
-});
 
 router.post('/votes/:id', async (req, res) => {
   const { id } = req.params;
