+++ test/e2e/moderationList.test.ts
@@ -2,10 +2,10 @@ import request from 'supertest';
 
 const HOST = `http://localhost:${process.env.PORT || 3000}`;
 
-describe('GET /moderationList', () => {
+describe('GET /api/moderationList', () => {
   describe('when fields params is empty', () => {
     it('returns all the list', async () => {
-      const response = await request(HOST).get('/moderationList');
+      const response = await request(HOST).get('/api/moderationList');
 
       expect(response.statusCode).toBe(200);
       expect(response.body).toMatchSnapshot();
@@ -16,7 +16,7 @@ describe('GET /moderationList', () => {
     it.each(['flaggedLinks', 'verifiedSpaces', 'flaggedProposals'])(
       'returns only the selected %s list',
       async field => {
-        const response = await request(HOST).get(`/moderationList?fields=${field}`);
+        const response = await request(HOST).get(`/api/moderationList?fields=${field}`);
 
         expect(response.statusCode).toBe(200);
         expect(response.body).toMatchSnapshot();
@@ -26,7 +26,7 @@ describe('GET /moderationList', () => {
 
   it('returns multiple list: verifiedSpaces,flaggedProposals', async () => {
     const response = await request(HOST).get(
-      `/moderationList?fields=verifiedSpaces,flaggedProposals`
+      `/api/moderationList?fields=verifiedSpaces,flaggedProposals`
     );
 
     expect(response.statusCode).toBe(200);
@@ -34,7 +34,9 @@ describe('GET /moderationList', () => {
   });
 
   it('ignores invalid field, and returns only verifiedSpaces', async () => {
-    const response = await request(HOST).get(`/moderationList?fields=verifiedSpaces,testInvalid`);
+    const response = await request(HOST).get(
+      `/api/moderationList?fields=verifiedSpaces,testInvalid`
+    );
 
     expect(response.statusCode).toBe(200);
     expect(response.body).toMatchSnapshot();
